Add helper functions for piece color and board parsing
